net_func_tests: Make use of TEST-NET range instead of future reserved range
IPv4 has, according to RFC 5737, three network ranges that are
specifically designed to be used for tests and code samples. Up
until now we were using 240.0.0.0/24 that is part of 240.0.0.0/4
which according to RFC 6890 is reserved for future use. Thus, it
is more appropriate to use one of the test networks and this patch
does just that.

Change-Id: I31ed725b3092da48723ea213ad80205961d9e65a
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/31259
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI