sudoers: Restrict usage of systemd-run
In commit 608e92e1bdd70 (utils: Add systemd_run command modifier) we
added systemd-run to the allowed commands in the sudoers configuration.
This change introduced a security breach, allowing vdsm to run any binary
using systemd-run.

This patch restricts usage of systemd-run to the mount command when
using glusterfs storage domain. Due to sudoers syntax, it is not
possible to allow any option in the systedmd-run command, and we must
specify the exact commend.

We also use systemd-run also for ifup and dhclinet, but these run from
supervdsm, so sudoers configuration is not needed.

This is an ugly fix for this security issue. A nicer fix requires more
work, calling the mount command in suprvdsm.

Change-Id: If39394371acd0f4c49433adf79638a8547bf81ec
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43937
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>