mount: Log info message when mounting or unmounting
Mounting and unmounting are important modifications to the machine state
and must be logged. Since actual mount command run now in supervdsm, we
have no clue about mounting and unmounting operations in vdsm log.

To help debugging issues with unresponsive mounts, we log also the time
it took to mount or unmount.

Change-Id: Idc89d6a1bfa0e9d117344d8870d81fed140d03c6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43974
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Adam Litke <alitke@redhat.com>