mount: More correct and simpler __eq__
Instance of different classes should not be considered equal, as
subclass may add new state not in the superclass.

If an instance is missing an attribute it should have, we don't want to
hide this error, and we certainly do not want to swallow *any* Exception
during the check.

This fixes the broken test_eq_subclass test.

Change-Id: Ib7b182fcd2ae95d2e0c310795a794f9e48d8d13f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43976
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>