mount: Better and simpler __hash__
The hash values were mixed in a poor way. There is no need to implement
this when we can use the hash function of the tuple object, which does a
better job.

For consistency, use self.__class__ for both equality test and computing
the hash. This is also more portable pattern, since type(self) returns
the same object for old style classes.

Change-Id: I6030c441c7137d65f4038d86e9813719429250c0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43978
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>