net: fix testSetupNetworksDeletesTheBridgeOnlyWhenItIsReconfigured
since the tap object itself went out of scope, the kernel removed
the device during the test since the file descriptor held by the
Tap object was closed. The test was also fixed to remove the device
before calling setupNetworks that would attempt to delete the bridge.

Change-Id: I52d0c04ffd3c047c1e66cec3d1a9646717d90a21
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43995
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>