lvm: Allow multiple initial tags when creating an LV
When creating new block volumes via the SDM VolumeArtifacts
infrastructure we want to create an LV with three tags to indicate: the
image, the parent volume, and that this volume is not yet completed.
Currently LVM only allows a single tag to be applied to new LVs.  Change
the behavior to accept a tuple of tags instead.

Change-Id: I2524f0127a7cb0eff5d7639b8ba1d916fd66af6a
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44051
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI