tests: Introduce indirection tests for storage refactoring
The refactoring of the storage classes to prepare for SPM removal is
moving around a lot of code but not changing the implementation.  The
SDM Indirection Tests check that calls made to the original objects
properly redirect to the implementation in the newly introduced classes.
To keep this refactoring simple, we are not testing the code itself,
just that it was moved correctly.

Change-Id: I336b543aa7ba019c91104e860a28bb40acc268b9
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44098
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI