tests: Fix error handling when checking dependencies
We used to hide *all* exceptions and skip the tests with possible bogus
message.

When trying brctl, we skip the test only if we cannot run it, since this
is the only error that can cause by missing bridge-utils. Any other
error should fail the tests.

When trying tc, we show the failure in the error message.

Also removed pointless code opening and closing /dev/null, which should
have been dropped in commit 61fff86c.

Change-Id: I180ba6449009dc986d2df58958039e2324d2191f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44110
Continuous-Integration: Jenkins CI
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>