net: always persist owned ifcfg files on ovirt node
Now, after VDSM writes all ifcfg files with ONBOOT=yes, and expecting
the corresponding networks to be up during boot, this needs to be
supported on ovirt-node. To do this, we need to make sure that all
owned ifcfg files are persisted during setSafeNetworkConfig so that they
will be there on next boot, for the network service to consider them.

Change-Id: Ibc717b86194a32c050d346e235a5c35fd318e1ff
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1203422
Reviewed-on: https://gerrit.ovirt.org/44133
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI