vm: make up pauseCode if libvirt is omitting it
When VDSM receives one IOError notification from libvirt,
it can override the VM pauseCode. The same pauseCode will
be reported to Engine through Events or through the
VM.getStats() API.

For historical reasons, libvirt reliably reports only the
'enospc' reason. In the other cases, the reason is an
empty string, and this resets the pauseCode to empty string.

Engine is not (anymore) expecting that.
To make Engine happy(er), and to improve the quality of data
VDSM reports in general, it is safe to report EOTHER
if the pauseCode is not given onIOError.

This patch does that.

Change-Id: I0bde0561cae20971effcf01c8a57159edcc4c42e
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44115
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI