tests: Fix ipwrapper tests isolation
Similar to tcTests, tests were sharing same devices, so failure to
delete a device would cause unrelated test to fail.

Even worse, the Unicode tests used *same* device name on all tests. So
running concurrent tests is not possible, and once we fail to delete a
device, all builds will fail on this slave when tyring to create the
same device again. This issue was hidden previously because we used to
skip silently such failures.

Now we use Unicode prefix with random suffix, so stale devices are
unlikely to cause failures in other builds.

To simplify cleanup, the Unicode tests was moved to its own test case
class. To make it easier to work with the Unicode bridge name, we use
plain ASCII ("\xd7\x90\xd7\x91\xd7\x92") instead of raw utf-8 ("אבג").

Change-Id: Ic02b3945a029476f17601608555430dc2fb710de
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44165
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>