tests: Remove hack for devices with same names
After the test isolation issues in tcTests.py and ipwrapperTests.py were
resolved, we are not expecting setup to fail because a device with same
name exists, so we can remove the hack that used to skip tests.

It is possible that slaves contains stale devices from previous run of
the broken tests, but this should be solved by cleaning the slaves, not
by hiding the error and skipping tests.

Change-Id: I675e57756859ca7ffe272607b7f528c281c65256
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44166
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>