Add confutil to manage properties files easily
In addition to confmeta [1], this util allows easy to manage files in
the format of

key = value
key2 = value
..

This uses configfile and add functionality for easy add and remove
sections inside the file.

Those function moved from libvirt configurator and are not introduced in
this patch. configurators/libvirt.py was the first usage until
abrt.conf. For later configurators this util might be useful.

See internal comments to understand usages.

[1] https://gerrit.ovirt.org/70583

Related-To: https://bugzilla.redhat.com/917062
Change-Id: Id67a0f51adb2d543c0542a48c020d34a13d1f2ed
Signed-off-by: Yeela Kaplan <ykaplan@redhat.com>
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>