nettestlib: Add requires_brctl and requries_tc decorators
Lot of tests used nettestlib.checkDependencies() in setUp() to skip all
tests if dependencies are not met. However, this function was checking
for two features, but most tests only need one of them, so tests could
be skipped for no reason.

Now we new specific requires decorators:

- @requires_brctl   raise SkipTest if running brctl fails
- @requires_tc      raise SkipTest if creating a bridge and adding a
                    qdisc fails.

nettestlib.checkDependencies() was creating a bridge and deleting it for
testing if we can run brctl. This could leave a stale bridge if deleting
the device failed. Now we use safer "brctl show".

Each test class is using now the specific requires decorator needed.

Change-Id: I9211cb491b1b6f5e7a8880559ec1e1ca3804e92a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44218
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI