caps: report capabilities on UNKNOWN OS
The capabilities reporting (getCapabilities API) does
logic on libvirt version to report, or not, the RNG source
availability.

The libvirt version is found using the package version,
but VDSM knows only how to interact with RPM and APT based
system, so it is clueless when it runs on UNKNOWN OS.

While it is still OK (from a VDSM PoV) to pass incomplete
information about available packages to Engine, this
makes the aforementioned RNG check logic fail.
Effects of this failure:
- Engine cannot connect to the VDS
- stacktrace in the VDSM logs (excerpt follows):

  File "/usr/share/vdsm/API.py", line 1314, in getCapabilities
    c = caps.get()
  File "/usr/share/vdsm/caps.py", line 665, in get
    '-'.join((caps['packages2']['libvirt']['version'],
KeyError: 'libvirt'
Thread-14::DEBUG::2015-07-31
11:16:52,699::stompreactor::305::yajsonrpc.StompServer::(send) Sending
response

This patch makes the getCapabilities API work again on
UNKNOWN OS. To be conservative, if we don't know about libvirt,
we report no RNG sources.

Please note that this prevents UNKONWN OS host to join cluster
with RNG sources configured.

Change-Id: Iaacb20dfd9b3850f477d8b45f221dc2ec186d0c2
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44232
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI