gluster: Allow "Replica" 1 domains
Patch 322ce98d17e585b156891ed711063d5e2c0ce5de introduced the ability to
verify the replica count when connecting to gluster storage, and set the
default allowed replica count to 3.

Contrary to the statement in that patch, "replica" 1 seems to be common
in the field, and we do not want to break (too many) existing
deployments.

This patch includes:
 - Changing the default configuration to support replica counts of 1 or
   3.
 - Update the description of this configuration value accordingly.
 - Update the test to match the default configuration and validate that
   having a value of '1,3' will indeed support replica counts of both 1
   and 3.

Bug-Url: https://bugzilla.redhat.com/1238093
Change-Id: Ia571d333e27357c3de825aaf29202cd4d1c4106a
Signed-off-by: Allon Mureinik <amureini@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44332
Continuous-Integration: Jenkins CI
Reviewed-by: Ala Hino <ahino@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>