hostdev: fix addressing
HOSTDEV:
Libvirt (domain_conf.c) does not expect the type for the host address,
and drops it if we send it. This causes USB address matching to fail
when parsing the domain XML.

SR-IOV
Interface device of type="hostdev" DO expect the address type as
it cannot be infered by hostdev's type.

Both versions expect 'type' in guest address.

Change-Id: I12592dd53f8af25fe8e1312b810e6b6f48411083
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44177
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI