recovery: move from clientIF to virt package
The code with recover VMs belongs to the virt vertical,
and was part of clientIF mostly for historical reasons.

To unclutter a bit the codebase, and to move forward the long
term goal to drop clientIF, this patch moves the vm recovery
code into a new virt module (virt/recovery.py).

Along the way, we also modernize the names and make them pep8
friendly. Besides that, there are no changes in logic.

Tests will follow up in a future patch.

Change-Id: I3c72782bb2e4a62f94514eb3059f2ba45f01b6e2
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44412
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>