storage: validate vdsm rwx permissions for file storage
VDSM needs permission to write to file-based storage.  Presumably the
existing check for qemu permissions or a subsequent failure would throw
an error; however, these errors were not always clear nor did they help
the user (e.g. a general Error 100, such as in the referenced bug).

With an explicit check, we can ensure the failure happens up-front and
gives a meaningful error message to the user.

Change-Id: I8276c36042bef57c80c8920041b1f3446cdfc372
Bug-Url: https://bugzilla.redhat.com/1250752
Signed-off-by: Greg Padgett <gpadgett@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44442
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>