storage: don't chmod local storage symlink
Proper permissions for the local storage target directory are already
verified before the local storage is connected, so there is no need to
change the permissions once the symlink has been created.

Additionally, there is no need to change the permissions of the symlink
itself.  From 'man 1 chmod':

  chmod never changes the permissions of symbolic links; the chmod
  system call cannot change their permissions.  This is not a problem
  since the permissions of symbolic links are never used.

The chmod call was added in 93390aab8381bff869bd029bf121f6a1f3761e8e in
a large patch with both refactoring changes and updates to the way
permissions were checked for storage.  However, the chmod is not needed
and is effectively redundant with the validation that runs when storage
is connected.

Change-Id: If1009033d7cb32add02c9bcc19840c601ea70b4e
Bug-Url: https://bugzilla.redhat.com/1112171
Signed-off-by: Greg Padgett <gpadgett@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44446
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Greg Padgett <gdpadgett@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/44990
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>