net: fix defaultRoute mismatch during restoration.
On rhev-h 3.5, when a host registers itself to the engine, the
management network is configured by vdsm-reg and VDSM internal
logic makes sure that this network will have DEFROUTE=yes in its
ifcfg file. Since VDSM logic happens during _addNetwork and after
the parameters that will eventually get persisted are set, we end
up with a network that is the default route but is not persisted
that way.
This tricks the selective restoration that think that this network
was somehow changed since it is different than its persistence. It
results in a restoration of the management network although it was
never changed.
This fix makes sure that if a management network is persisted, it
will always be considered as defaultRoute.
Since VDSM 3.6, vdsm registration along with internal logic are
gone, and all defaultRoute logic lies in engine. Hence this patch
should only be part of 3.5.

Label: ovirt-3.5-only
Change-Id: If8e4d19329b230d620bc4545a0d0355f482b5bef
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1203422
Reviewed-on: https://gerrit.ovirt.org/44613
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Continuous-Integration: Yaniv Bronhaim <ybronhei@redhat.com>