network: add ifcfg write hook
Added a new hook point which is executed before and after
an ifcfg file is writen on the host as a result of
executing setupNetworks.
An example can be a use case where the user would like
to tweak the vdsm-produced ifcfg files by adding
PEERDNS=no to it.
The hook script receives a json file containing the
following dict:
name: the name of the interface being configured
ifcfg_file: full path of the interface file being written
config: the contents of the config file being written

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=965929
Change-Id: I2206ccf97b210bd58f03777d4ccd016785b02939
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44552
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI