network: bridge inherits DHCP unique identifier from its DHCP-enabled port
On Fedora 21 or newer, dhclient has started to use DHCP unique identifier
with a link-layer address and time (DUID-LLT) to identify a host.

When a host (e.g. oVirt node) is being added to engine's management network
while its NIC used DHCP, running dhclient on the bridge on top of the NIC
results in connectivity loss because the new dhclient is unaware of the DUID
used previously (dhclient should store it somewhere; bug to be reported soon)
and so the newly-generated DUID contains a different time value (albeit the
bridge's MAC address is the same as the NIC's), thus the host is given
a different IP address as well.

This patch assures that the original DUID is read from the NIC's dhclient
lease file by the dhclient running on the bridge for all network
configurators. A new functional test (a specialized copy of
testSetupNetworksAddDelDhcp) verifies the fix.

Change-Id: If91b46657be8a5a42a118b23155d059df379727e
Bug-Url: https://bugzilla.redhat.com/1219429
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44539
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI