net: create bond properly during start_devices
This method, that is used during networks rollback, calls ifup on
all ifcfg files it knows of. However, to ifup a bond, one needs to
manually create it first, by writing '+bondingName' into
/sys/class/net/bonding_masters . Ifcfg configurators already does
so when creating bonds. Skipping this step causes ifup to fail and
complain: 'Device bond11 does not seem to be present, delaying
initialization.'
This also fixes
testSetupNetworksEmergencyDevicesCleanupBondOverwrite which is
almost never run because it is a slow test.

Change-Id: Ic48f3e8b51e4e7e8668902a785c7c271ada03df6
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45027
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>