supervdsm: Decorator for supervdsm proxied calls.
These decorators allow automatic dispatch to supervdsm process, in this
way:

@supervdsm.proxied_call
def some_func():
	this_requires_root()

Calling the above function from root context (such as supervdsm) will
execute it as regular code, but when calling from non-root context,
it will try and dispatch it to supervdsm process.

Inside the supervdsm process, it will attempt to import the requested
module and retrieve the relevant function, then call it and return the
result.

Change-Id: Ibea2b07b91a7faffda16c71c6b5831f96d116c15
Signed-off-by: Dima Kuznetsov <dkuznets@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44598
Continuous-Integration: Jenkins CI
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>