vm: make acpiShutdown handle NotConnectedError
In the Vm shutdown flow it is possible that the domain is
disconnected asynchronously as in any other Vm flow.

This patch makes Vm.acpiShutdown() handle the
sudden domain disconnections in a graceful manner.

vmpowerdown needs to be aware if acpiShutdown failed
for this reason, so it can go ahead with the callback
chain without wasteful waits,

Change-Id: I244f00d62ee24fb42ba3d654961a8fc22f4a6c25
Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1154389
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/34879
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45379
Reviewed-by: Martin Polednik <mpolednik@redhat.com>