vm: Libvirt quering after disk detach operation addition.
As stated in libvirt documentary, after detaching a device using
virDomainDetachDeviceFlags, we need to verify that this device
has actually been detached.

Currently we use virDomainDetachDevice. However- That function behaves
the same in that matter. (Currently it is not documented at libvirt's
API docs- but after contacting libvirt's guys it turned out that this
is true. Bug 1257280 opened for fixing the documentation.)

Not verifying that the device was detached, as mentioned above, cause
various problems, as hotunplugDisk could return a success result
while it did not actually succeeds to detach the disk.

This patch adds this functionallity to hotunplugDisk, and after some
timeout fails the operation if the disk was not detached.

Change-Id: I393ce55dd761ac825cb96bd499976fd74c366b09
Bug-Url: https://bugzilla.redhat.com/1044466
Signed-off-by: Amit Aviram <aaviram@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45138
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45646
Reviewed-on: https://gerrit.ovirt.org/47629
Continuous-Integration: Francesco Romani <fromani@redhat.com>