debug: Make it easy to debug start failures
If Vdsm fail to start before logging was configured, it used to log a
message to syslog and exit silently. This make it hard to debug from the
shell.

This patch raises FatalError in the places that used to log to syslog
and exit.  In the exception handler we log to syslog and re-raise the
exception, so the traceback is available in the shell, or in systemd
journal.

For example, instead of:

    # /usr/share/vdsm/vdsm
    (no output)

We get now:

    # /usr/share/vdsm/vdsm
    Traceback (most recent call last):
      File "/usr/share/vdsm/vdsm", line 271, in <module>
        main()
      File "/usr/share/vdsm/vdsm", line 253, in main
        __assertVdsmUser()
      File "/usr/share/vdsm/vdsm", line 230, in __assertVdsmUser
        % (constants.VDSM_USER, username))
    __main__.FatalError: Not running as 'vdsm', trying to run as 'root'

In the system log we will see this message:

    Aug 11 20:58:34 hostname vdsm: VDSM failed to start: Not running as
    'vdsm', trying to run as 'root'

Change-Id: I7c47e289ce4d7d043d8d95887a3d645797270944
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44728
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>