vmstats: network: avoid ZeroDivisionError
For still under investigation case, after migration
sometimes vmstats code is fed with zero-intervalled samples.
While we investigate the root cause, handle gracefully ZeroDivisionError
adding guards in the public functions that have an 'interval' argument.

The only left one was 'networks', so this patch fixes that.

It is not good enough to generalize the check in vmstats.produce()
because the relevant functions (cpu, disks, networks) are part of
the public API of the module, so they should handle bad input properly.

To improve testability, we fix the return value of the aforementioned
functions. The return value is meant to be used only in testing.

X-Backport-To: 3.6
Change-Id: I1ff99e3706212179ef9dc56e92e7d5decdd6bd7a
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44764
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>