supervdsm: generalize udevTrigger method
udevTrigger in it's current state is tied to storage code as it does
property match against DM_NAME. USB devices require attribute match for
devnum and busnum, and we should be somewhat able to do both these
matches in a more generic function.

This patch refactors udevTrigger to include attr_matches and
property_matches. When an iterable of attribute-value iterables is sent,
they're expanded to the correct udevadm command.

Change-Id: I761305a54c14e8ac8ef9f4204ee5cc5d55cb31b7
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44767
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>