virt: periodic: reduce NotConnectedError noise
The periodic operations have known races on
VM startup and on VM shutdown.

These races are because periodic are asynchronous with
respect to the VM startup/shutdown, and they are benign
because operations will just be always retried again in
the next periodic cycle.

This is particularly visible on loaded servers and/or
slowly-responsive libvirt.

Backport-To: 3.6
Change-Id: I920e3b0b0e80e0a66ad199607068424986933d3a
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44813
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>