storage: Cleanup sd.getVAllocSize redirection
For file-based storage, getVSize and getVAllocSize have different
implementations.  For block storage, getVAllocSize just calls getVSize.
The logic stating that the implementation of these functions is the same
on block storage should be part of the BlockStorageDomainManifest not
the BlockStorageDomain.  Therefore, we create getVAllocSize in the
BlockStorageDomainManifest and have it call its own getVSize.  Now block
and file StorageDomains can share the same indirection method for
getVAllocSize (just as they do for getVSize).

Change-Id: I55a277caeee1747e8c9e2869a557d3b97fa767ec
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44829
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI