vm: graphics: fix settings after restore
When restoring a VM from a memory snapshot,
if we don't reset the some details of the SPICE configuration
will be lost, and the restored VM will behave erratically.

This behaviour is easily reproduced by:
a. boot a VM with SPICE graphic device
b. do a snapshot, make sure to check 'snapshot memory'
c. preview the snapshot
d. run the VM
e. check the VM parameters (e.g. qemu)

The automatic restore using virDomainRestoreFlags() uses the wrong
configuration settings.

To fix this, we need to amend the VM Domain XML, tuning the configuration
the same way we do for VM startup.

Backport-To: 3.6
Backport-To: 3.5
Bug-Url: https://bugzilla.redhat.com/1260177
Change-Id: Ic9f11e2d87cc715c77d305c005c1cd7f502d506d
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44842
Continuous-Integration: Jenkins CI
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46524
Reviewed-on: https://gerrit.ovirt.org/47487