vm: fix misleading XML log
When VDSM is asked to create a new VM, among the other things,
it builds the domain XML according to the configuration sent
by client (aka Engine) and run the before_vm_start hook.

This is done also on the dehibernating (aka migration from file)
path, for no apparent good reason.

This is wasteful and useless at best, and most likely misleading,
because VDSM logs the XML it produced, which is *not* what is
gonna be used in the dehibernation path.

This patch fixes this issue loggging the XML which is actually
about to be fed to libvirt, not the useless one regenerated
by VDSM.

Backport-To: 3.6
Change-Id: I50f5b8a51830c8ce2a098e4ac11dd1bd6a89aac3
Bug-Url: https://bugzilla.redhat.com/1292371
Related-To: https://bugzilla.redhat.com/1190210
Related-To: https://bugzilla.redhat.com/1234197
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44843
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49139
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>