supervdsm: move udevadm methods to udevadm module
udevadm module already exists, so there is no reason to keep the
method logic in supervdsm. udevadm also provides us with a means to
execute a command, it is therefore used instead of EXT_UDEVADM. To
keep the behavior same, Error raised by trigger is caught in
supervdsm and re-raised as IOError (same as before the patch).

Change-Id: Ie2b72b580e77872068ae6357410d940a83626b05
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44808
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>