multipath: Reformat old tags list
It is easier to work with long lists when we have one item per line, and
all lines are terminated by ",". This is not possible in some languages,
but Python wisely allow this syntax. Adding a new tag is now one line
change and the diffs are nicer.

Change-Id: I45f24a1c90874c72b24fcd41006786e3822a4379
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44862
Continuous-Integration: Jenkins CI
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>