multiapth: Reorganize constants and improve documentation
It is nicer to work with code when related items are grouped together
and when groups have documentation.

Now the module start with a description of the VDSM REVISION tag, which
must be in the first line of the file, followed by the current and old
tag values.

Then we discuss the VDSM PRIVATE tag, which may be in the second line,
followed by the current and old tag values.

Previously you had to read the code to understand this logic.

Change-Id: Id853b50ede32d9f0b05ef610ec5008b1d4ead2da
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44863
Continuous-Integration: Jenkins CI
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>