multipath: Write multipath.conf atomically
We used to create a temporary file, and then copy (using cp) the
temporary file to final location.  Using external process to copy files
on local storage is not needed, we can use the Python standard library
for that. Also, this operation is not atomic, and can leave half written
file if the write is interrupted in the middle.

Now we use NamedTemporaryFile(delete=False), to create a temporary file
safely, but do not delete it, so we can rename the file atomically.

Change-Id: I850d621b7cb09f2732b8b3eb2cb2897e87547ddb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44869
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>