net: always persist ifcfg files.
Currently the persistence of ifcfg files is only done during
setSafeNetworkConfig. But, if the user somehow opts out of this (by
unticking this option in engine, or if setupNetwork is called out of
this flow (e.g. vdsm-restore-net-config) then vdsm unpersists the
ifcfg files. If then the host reboots without persisting the
unpersisted ifcfg files, those files are gone during boot since they
are not part of the original file system.

Change-Id: I02ae28c345111309536e4af2bd6a09799f38b82f
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1252268
Reviewed-on: https://gerrit.ovirt.org/44926
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>