virt: add and use vm.acpiReboot() method
The VM powerdown flow, implemented into the vmpowerdown.py module,
needs to call libvirt APIs to trigger the actual shutdown or reboot
operations.

To do so, it directly accesses the Vm._dom private attribute, which
is impolite and evil.

This patch adds acpiReboot public method, with built-in protection
against asynchronous QEMU death.
This of course adds more code to the already fat vm.py, but improves
isolation among modules, which is a greater good.

Change-Id: I5d6dd3a81cd4ad981a839849a4a2914e35a53fa6
Related-To: https://bugzilla.redhat.com/1154389
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44923
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>