multipath: Remove unneeded "features" configuration
We used to override "features 0" to disable "queue_if_no_path" feature.
However, this configuration also disable other unrelated features, that
may be needed for optimal configuration.

We use "no_path_retry fail", which removes the feature
"queue_if_no_path" if it exists in the "features" option.

For example, if a built-in device is using:

    features "3 queue_if_no_path pg_init_retries 50"

When overriding "no_path_retry fail", the device features becomes:

    features "2 pg_init_retries 50"

Now we override only "no_path_retry fail" and leave "features" as is.

Change-Id: I6c77c5a3574e3dd7ec2f740e194a6e59efc46c2a
Backport-To: 3.6
Relates-To: https://bugzilla.redhat.com/880738
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44943
Continuous-Integration: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>