hsm: Support checkStatus param in getDeviceList
In order to populate the 'status' field, the getDeviceList
verb perform a create PV test. This operation is expensive, and in
setups with large number of devices it will cause performance issue.

However, the 'status' field is not always needed. This patch
add an optional 'checkStatus' boolean parameter to getDeviceList.
By default it will be True to keep same behavior as before.

If specified as False, the PV create test will be skipped and the
'status' field will be populated as 'unknown'.

The flow before this patch was:
    - Engine calls getDeviceList, VDSM perform status check for each
      devices, possibly times out
    - User select a few devices (typically one)
    - Engine warn user about used devices

The flow with this patch:
    - Engine calls getDeviceList, skipping status check
    - User select a few devices (typically one)
    - Engine calls getDeviceList, checking status only for selected
    - Engine warn user about used devices

On a setup of 100 ISCSI devices, the command took 3s without
the PV create test against 58s with the test.

real    0m2.987s
user    0m0.263s
sys     0m0.028s

real    0m57.769s
user    0m0.265s
sys     0m0.024s

Examples:
    getDeviceList
        return all devices
    getDeviceList FCP
        return only FCP devices
    getDeviceList FCP True
        return only FCP devices and perform PV create test
    getDeviceList ISCSI False guid1 guid2
       return info for guid1 and guid2, assuming ISCSI type
       without performing PV create test

Change-Id: Ic28954708f2fd7c7b721aa7f9a0fb6e1a6019597
Bug-Url: https://bugzilla.redhat.com/1217401
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45093
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>